requirements/base.txt
django==4.2.9
python-dotenv==1.0.1
djangorestframework==3.15.2
setuptools==75.8.0
django-cors-headers==3.10.1
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
drf-yasg==1.20.0
Pillow==10.4.0
argon2-cffi==21.3.0
pytz==2025.1
redis==5.0.3
django-redis==5.4.0
celery==5.3.6
flower==2.0.0
django-celery-email==3.0.0
djoser==2.3.1
django-allauth==65.0.0
dj-rest-auth==7.0.1
djangorestframework-simplejwt==5.3.0
PyJWT==2.10.0
django-haystack==3.1.1
Whoosh==2.7.4
drf-haystack==1.9.1
argon2-cffi==21.2.0